Moved Velocity tests to JUnit 5
